created a PEEP_Base to remove duplicate code. also changed the __init__ file to run local tests
