+++ netsec_fall2017/lab2/src/Peep_Test.py
@@ -2,7 +2,6 @@ import asyncio
 import playground
 import sys
 import lab2_protocol
-from playground.network.common import StackingProtocolFactory
 from RNG_game_protocol import RandomNumberGameServerProtocol, RandomNumberGameClientProtocol
 
 USAGE = """usage: Peep_Test <mode>
@@ -15,7 +14,7 @@ if __name__ == "__main__":
     elif args[0] == 'server':
         loop = asyncio.get_event_loop()
         loop.set_debug(enabled=True)
-        coro = playground.getConnector("lab2_protocol").create_playground_server(RandomNumberGameServerProtocol, 8888)
+        coro = playground.getConnector("lab2_protocoltest").create_playground_server(RandomNumberGameServerProtocol, 8888)
 
         server = loop.run_until_complete(coro)
 
@@ -31,7 +30,7 @@ if __name__ == "__main__":
     else:
         loop = asyncio.get_event_loop()
         loop.set_debug(enabled=True)
-        coro = playground.getConnector("lab2_protocol").create_playground_connection(lambda: RandomNumberGameClientProtocol(loop), args[0], 8888)
+        coro = playground.getConnector("lab2_protocoltest").create_playground_connection(lambda: RandomNumberGameClientProtocol(loop), args[0], 8888)
 
         socket, client_proto = loop.run_until_complete(coro)
         loop.run_forever()
