+++ netsec_fall2017/lab2/src/lab2_protocol/__init__.py
@@ -2,4 +2,4 @@ import playground
 from .Peep_Passthrough import clientFactory, serverFactory
 
 lab2Connector = playground.Connector(protocolStack = (clientFactory, serverFactory))
-playground.setConnector("lab2_protocol", lab2Connector)
+playground.setConnector("lab2_protocoltest", lab2Connector)
